created main module, refactored execution module
