+++ execution/src/main/java/executor/service/execution/exception/step/ClickCssException.java
@@ -1,4 +1,4 @@
-package executor.service.exception.step;
+package executor.service.execution.exception.step;
 
 
 public class ClickCssException extends StepExecutionException {
