+++ execution/src/main/java/executor/service/execution/exception/step/UnknownStepException.java
@@ -1,4 +1,4 @@
-package executor.service.exception.step;
+package executor.service.execution.exception.step;
 
 public class UnknownStepException extends StepExecutionException {
     public UnknownStepException(String message) {
