+++ execution/src/main/java/executor/service/execution/model/ThreadPoolConfig.java
@@ -1,9 +1,10 @@
-package executor.service.model;
+package executor.service.execution.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Component;
 
 
@@ -11,6 +12,7 @@ import org.springframework.stereotype.Component;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+@PropertySource("classpath:execution.properties")
 public class ThreadPoolConfig {
     @Value("${executorservice.common.threadsCount}")
     private Integer corePoolSize;
