+++ execution/src/main/java/executor/service/execution/scenario/step/ClickCss.java
@@ -1,7 +1,7 @@
 package executor.service.execution.scenario.step;
 
 import executor.service.logger.annotation.Logged;
-import executor.service.exception.step.ClickCssException;
+import executor.service.execution.exception.step.ClickCssException;
 import executor.service.model.Step;
 import org.openqa.selenium.*;
 import org.springframework.stereotype.Component;
