+++ execution/src/main/java/executor/service/execution/scenario/step/ClickXpath.java
@@ -1,7 +1,7 @@
 package executor.service.execution.scenario.step;
 
 import executor.service.logger.annotation.Logged;
-import executor.service.exception.step.ClickXPathException;
+import executor.service.execution.exception.step.ClickXPathException;
 import executor.service.model.Step;
 import org.openqa.selenium.*;
 import org.springframework.stereotype.Component;
