+++ execution/src/main/java/executor/service/execution/scenario/step/Sleep.java
@@ -1,7 +1,7 @@
 package executor.service.execution.scenario.step;
 
 import executor.service.logger.annotation.Logged;
-import executor.service.exception.step.SleepException;
+import executor.service.execution.exception.step.SleepException;
 import executor.service.model.Step;
 import org.openqa.selenium.WebDriver;
 import org.springframework.stereotype.Component;
