+++ execution/src/test/java/executor/service/execution/scenario/ScenarioExecutorImplTest.java
@@ -1,7 +1,7 @@
 package executor.service.execution.scenario;
 
-import executor.service.exception.SiteNotFoundException;
-import executor.service.exception.step.UnknownStepException;
+import executor.service.execution.exception.SiteNotFoundException;
+import executor.service.execution.exception.step.UnknownStepException;
 import executor.service.execution.scenario.step.StepExecution;
 import executor.service.model.Scenario;
 import executor.service.model.Step;
