+++ settings.gradle.kts
@@ -6,3 +6,4 @@ include("model")
 include("webdriver")
 include("collection")
 include("source")
+include("main")
